moved reminder_needed? methods into CalendarHelper
