+++ app/helpers/medication_refill_helper.rb
@@ -1,6 +1,7 @@
 # frozen_string_literal: true
 
 module MedicationRefillHelper
+  include CalendarHelper
   # Save refill date to Google calendar
   def save_refill_to_google_calendar(medication)
     return true unless current_user.google_oauth2_enabled? &&
@@ -20,25 +21,5 @@ module MedicationRefillHelper
       date: medication.refill,
       access_token: current_user.google_access_token,
       email: current_user.email }
-  end
-
-  def new_cal_refill_reminder_needed?(medication)
-    if medication.add_to_google_cal &&
-       medication.refill &&
-       (medication.add_to_google_cal_changed? || medication.refill_changed?)
-      true
-    else
-      false
-    end
-  end
-
-  def new_cal_refill_reminder_unneeded?(medication)
-    if medication.add_to_google_cal &&
-       medication.refill &&
-       (medication.add_to_google_cal_changed? || medication.refill_changed?)
-      false
-    else
-      true
-    end
   end
 end
